+++ packages/shared_preferences/shared_preferences_android/lib/shared_preferences_android.dart
@@ -19,6 +19,8 @@ class SharedPreferencesAndroid extends SharedPreferencesStorePlatform {
     SharedPreferencesStorePlatform.instance = SharedPreferencesAndroid();
   }
 
+  static const String _defaultPrefix = 'flutter.';
+
   @override
   Future<bool> remove(String key) async {
     return (await _kChannel.invokeMethod<bool>(
@@ -37,17 +39,28 @@ class SharedPreferencesAndroid extends SharedPreferencesStorePlatform {
 
   @override
   Future<bool> clear() async {
-    return (await _kChannel.invokeMethod<bool>('clear'))!;
+    return clearWithPrefix(_defaultPrefix);
+  }
+
+  @override
+  Future<bool> clearWithPrefix(String prefix) async {
+    return (await _kChannel.invokeMethod<bool>(
+      'clearWithPrefix',
+      <String, dynamic>{'prefix': prefix},
+    ))!;
   }
 
   @override
   Future<Map<String, Object>> getAll() async {
-    final Map<String, Object>? preferences =
-        await _kChannel.invokeMapMethod<String, Object>('getAll');
+    return getAllWithPrefix(_defaultPrefix);
+  }
 
-    if (preferences == null) {
-      return <String, Object>{};
-    }
-    return preferences;
+  @override
+  Future<Map<String, Object>> getAllWithPrefix(String prefix) async {
+    return (await _kChannel.invokeMapMethod<String, Object>(
+          'getAllWithPrefix',
+          <String, dynamic>{'prefix': prefix},
+        )) ??
+        <String, Object>{};
   }
 }