+++ packages/webview_flutter/webview_flutter_android/android/src/test/java/io/flutter/plugins/webviewflutter/WebViewTest.java
@@ -10,17 +10,22 @@ import static org.junit.Assert.assertTrue;
 import static org.mockito.ArgumentMatchers.any;
 import static org.mockito.ArgumentMatchers.eq;
 import static org.mockito.Mockito.mock;
+import static org.mockito.Mockito.spy;
 import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.when;
 
 import android.content.Context;
+import android.os.Build;
+import android.view.View;
 import android.webkit.DownloadListener;
 import android.webkit.ValueCallback;
 import android.webkit.WebChromeClient;
 import android.webkit.WebViewClient;
+import io.flutter.embedding.android.FlutterView;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugins.webviewflutter.GeneratedAndroidWebView.WebViewFlutterApi;
 import io.flutter.plugins.webviewflutter.WebViewHostApiImpl.WebViewPlatformView;
+import io.flutter.plugins.webviewflutter.utils.TestUtils;
 import java.util.HashMap;
 import java.util.Objects;
 import org.junit.After;
@@ -54,11 +59,7 @@ public class WebViewTest {
         .thenReturn(mockWebView);
     testHostApiImpl =
         new WebViewHostApiImpl(
-            testInstanceManager,
-            mockBinaryMessenger,
-            mockWebViewProxy,
-            mockContext,
-            null);
+            testInstanceManager, mockBinaryMessenger, mockWebViewProxy, mockContext, null);
     testHostApiImpl.create(0L, true);
   }
 
@@ -314,9 +315,9 @@ public class WebViewTest {
           }
         };
 
-    testInstanceManager.addDartCreatedInstance(webView, 0);
+    testInstanceManager.addDartCreatedInstance(webView, 1);
     final JavaObjectHostApiImpl javaObjectHostApi = new JavaObjectHostApiImpl(testInstanceManager);
-    javaObjectHostApi.dispose(0L);
+    javaObjectHostApi.dispose(1L);
 
     assertTrue(destroyCalled[0]);
   }
@@ -341,11 +342,26 @@ public class WebViewTest {
   }
 
   @Test
+  public void setImportantForAutofillForParentFlutterView() {
+    final WebViewPlatformView webView =
+        new WebViewPlatformView(mockContext, mockBinaryMessenger, testInstanceManager);
+
+    final WebViewPlatformView webViewSpy = spy(webView);
+    final FlutterView mockFlutterView = mock(FlutterView.class);
+    when(webViewSpy.getParent()).thenReturn(mockFlutterView);
+
+    TestUtils.setFinalStatic(Build.VERSION.class, "SDK_INT", Build.VERSION_CODES.O);
+    webViewSpy.onAttachedToWindow();
+
+    verify(mockFlutterView).setImportantForAutofill(View.IMPORTANT_FOR_AUTOFILL_YES);
+  }
+
+  @Test
   public void onScrollPosChange() {
     final InstanceManager instanceManager = InstanceManager.open(identifier -> {});
 
     final WebViewFlutterApiImpl flutterApiImpl =
-        new WebViewFlutterApiImpl(mockBinaryMessenger, instanceManager);
+            new WebViewFlutterApiImpl(mockBinaryMessenger, instanceManager);
 
     final WebViewFlutterApi mockFlutterApi = mock(WebViewFlutterApi.class);
     flutterApiImpl.setApi(mockFlutterApi);
@@ -354,9 +370,9 @@ public class WebViewTest {
     flutterApiImpl.onScrollPosChange(mockWebView, 0L, 1L, 2L, 3L, reply -> {});
 
     final long instanceIdentifier =
-        Objects.requireNonNull(instanceManager.getIdentifierForStrongReference(mockWebView));
+            Objects.requireNonNull(instanceManager.getIdentifierForStrongReference(mockWebView));
     verify(mockFlutterApi)
-        .onScrollPosChange(eq(instanceIdentifier), eq(0L), eq(1L), eq(2L), eq(3L), any());
+            .onScrollPosChange(eq(instanceIdentifier), eq(0L), eq(1L), eq(2L), eq(3L), any());
 
     instanceManager.close();
   }