+++ packages/webview_flutter/webview_flutter_android/lib/src/android_webview_api_impls.dart
@@ -933,8 +933,16 @@ class WebStorageHostApiImpl extends WebStorageHostApi {
 /// Flutter api implementation for [FileChooserParams].
 class FileChooserParamsFlutterApiImpl extends FileChooserParamsFlutterApi {
   /// Constructs a [FileChooserParamsFlutterApiImpl].
-  FileChooserParamsFlutterApiImpl({InstanceManager? instanceManager})
-      : instanceManager = instanceManager ?? JavaObject.globalInstanceManager;
+  FileChooserParamsFlutterApiImpl({
+    this.binaryMessenger,
+    InstanceManager? instanceManager,
+  }) : instanceManager = instanceManager ?? JavaObject.globalInstanceManager;
+
+  /// Receives binary data across the Flutter platform barrier.
+  ///
+  /// If it is null, the default BinaryMessenger will be used which routes to
+  /// the host platform.
+  final BinaryMessenger? binaryMessenger;
 
   /// Maintains instances stored to communicate with java objects.
   final InstanceManager instanceManager;
@@ -953,6 +961,8 @@ class FileChooserParamsFlutterApiImpl extends FileChooserParamsFlutterApi {
         acceptTypes: acceptTypes.cast(),
         mode: mode.value,
         filenameHint: filenameHint,
+        binaryMessenger: binaryMessenger,
+        instanceManager: instanceManager,
       ),
       instanceId,
     );