+++ packages/webview_flutter/webview_flutter_android/test/android_navigation_delegate_test.dart
@@ -5,13 +5,23 @@
 import 'dart:async';
 
 import 'package:flutter_test/flutter_test.dart';
+import 'package:mockito/annotations.dart';
 import 'package:webview_flutter_android/src/android_proxy.dart';
 import 'package:webview_flutter_android/src/android_webview.dart'
     as android_webview;
 import 'package:webview_flutter_android/webview_flutter_android.dart';
 import 'package:webview_flutter_platform_interface/webview_flutter_platform_interface.dart';
 
+import 'android_navigation_delegate_test.mocks.dart';
+import 'test_android_webview.g.dart';
+
+@GenerateMocks(<Type>[TestInstanceManagerHostApi])
 void main() {
+  TestWidgetsFlutterBinding.ensureInitialized();
+
+  // Mocks the call to clear the native InstanceManager.
+  TestInstanceManagerHostApi.setup(MockTestInstanceManagerHostApi());
+
   group('AndroidNavigationDelegate', () {
     test('onPageFinished', () {
       final AndroidNavigationDelegate androidNavigationDelegate =
@@ -452,6 +462,7 @@ AndroidNavigationDelegateCreationParams _buildCreationParams() {
 }
 
 // Records the last created instance of itself.
+// ignore: must_be_immutable
 class CapturingWebViewClient extends android_webview.WebViewClient {
   CapturingWebViewClient({
     super.onPageFinished,
@@ -460,6 +471,8 @@ class CapturingWebViewClient extends android_webview.WebViewClient {
     super.onReceivedRequestError,
     super.requestLoading,
     super.urlLoading,
+    super.binaryMessenger,
+    super.instanceManager,
   }) : super.detached() {
     lastCreatedDelegate = this;
   }
@@ -480,6 +493,8 @@ class CapturingWebChromeClient extends android_webview.WebChromeClient {
   CapturingWebChromeClient({
     super.onProgressChanged,
     super.onShowFileChooser,
+    super.binaryMessenger,
+    super.instanceManager,
   }) : super.detached() {
     lastCreatedDelegate = this;
   }
@@ -491,6 +506,8 @@ class CapturingWebChromeClient extends android_webview.WebChromeClient {
 class CapturingDownloadListener extends android_webview.DownloadListener {
   CapturingDownloadListener({
     required super.onDownloadStart,
+    super.binaryMessenger,
+    super.instanceManager,
   }) : super.detached() {
     lastCreatedListener = this;
   }