+++ packages/webview_flutter/webview_flutter_android/CHANGELOG.md
@@ -1,6 +1,15 @@
-## 3.4.3
+## 3.4.5
 
 * Added `setOnContentOffsetChanged` method to the `AndroidWebViewController`.
+
+## 3.4.4
+
+* Fixes a bug where the native `WebView` wouldn't be traversed for autofill automatically.
+* Updates minimum Flutter version to 3.3.
+
+## 3.4.3
+
+* Updates internal Java InstanceManager to be cleared on hot restart.
 
 ## 3.4.2